Add files via upload
This code block demonstrates the functionality of the API, and will later be integrated with the full project. I dedicated the second week to getting the API working because I wanted to make sure that there would not be problems with it down the line once the rest of the project was already completed.